Backend restructuring, moved generic code out of api and into webviz-services
