+++ webviz-services/webviz_services/utils/authenticated_user.py
